Fix issue related to function parameters
It was a big one, and lots of things learned here. My problem involved
passing the task array to different functions. They used the
"starting" copy of it. This means that content on the page did not
update to reflect the current state of affairs. In the end, I managed
to get over the bug by passing a function that returned the task
array value in these functions. Yay!